move get_updated_datasets
